services/discord-bot/requirements.txt
discord.py>=2.3.0
docker>=6.0.0
psutil>=5.9.0
